search, filter, pagination and group added.

+++ Controllers/StudentsController.cs
@@ -21,9 +21,66 @@ namespace HertfordshireUniversity.Controllers
         }
 
         // GET: Students
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(
+            string sortOrder, string searchString, string currentFilter, int? pageNumber)
         {
-            return View(await _context.Students.ToListAsync());
+            ViewData["CurrentSort"] = sortOrder;
+            ViewData["NameSortParm"] = string.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+            ViewData["DateSortParm"] = sortOrder == "Date" ? "date_Desc" : "Date";
+            ViewData["FirstNameSortParm"] = sortOrder == "First_Asc" ? "First_Desc" : "First_Asc";
+
+            if(searchString != null)
+            {
+                pageNumber = 1;
+            }
+            else
+            {
+                searchString = currentFilter;
+            }
+
+            ViewData["CurrentFilter"] = searchString;
+
+            var students = from s in _context.Students
+                           select s;
+
+            if (!string.IsNullOrEmpty(searchString))
+            {
+                students = students.Where(x => x.LastName.Contains(searchString) ||
+                                             x.FirstMidName.Contains(searchString));
+            }
+
+            switch (sortOrder)
+            {
+                case "name_desc":
+                    students = students.OrderByDescending(x => x.LastName);
+                    break;
+
+                case "Date":
+                    students = students.OrderBy(x => x.EnrollmentDate);
+                    break;
+
+                case "date_Desc":
+                    students = students.OrderByDescending(x => x.EnrollmentDate);
+                    break;
+
+                case "First_Asc":
+                    students = students.OrderBy(x => x.FirstMidName);
+                    break;
+
+                case "First_Desc":
+                    students = students.OrderByDescending(x => x.FirstMidName);
+                    break;
+
+                default:
+                    students = students.OrderBy(x => x.LastName);
+                    break;
+
+            }
+
+            int pageSize = 3;
+            return View(await PaginatedList<Student>.CreateAsync(students.AsNoTracking(),pageNumber??1,pageSize));
+
+            //return View(await students.AsNoTracking().ToListAsync());
         }
 
         // GET: Students/Details/5
@@ -58,7 +115,7 @@ namespace HertfordshireUniversity.Controllers
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
 
-        public StudentViewModel studentViewModel { get; set; }
+        //public StudentViewModel studentViewModel { get; set; }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -110,7 +167,7 @@ namespace HertfordshireUniversity.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != null)
+            if (id == null)
             {
                 return NotFound();
             }
@@ -141,7 +198,7 @@ namespace HertfordshireUniversity.Controllers
 
 
         // GET: Students/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -149,10 +206,19 @@ namespace HertfordshireUniversity.Controllers
             }
 
             var student = await _context.Students
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
+
             if (student == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(student);
@@ -164,9 +230,23 @@ namespace HertfordshireUniversity.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var student = await _context.Students.FindAsync(id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+            if (student == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Students.Remove(student);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool StudentExists(int id)
